Moved 1D,2D,3D dcf with  voronoi calculations to algorithms
